+++ cli/cli/CHANGELOG.md
@@ -10,10 +10,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - `--unmask-logs` option will show full tokens in verbose logs
 - `--no-log-file` option will prevent verbose logs from being written to temp file
-
-### Added
 - `beam project enable` and `beam project disable` commands will set the `<BeamEnabled>` setting.
 - Can pass MSBuild dlls location through environment variable to the CLI
+- `beam services build` uses Docker Buildkit to build Standalone Microservice images
+- `beam services bundle` produces a `.tar` file for a Standalone Microservice
+- `--docker-cli-path` option overrides docker cli location used for Buildkit
+- `net8.0` support for Standalone Microservices
+
+### Changed
+- Standalone Microservices are created with `net8.0` by default
 
 
 ## [2.0.1] - 2024-06-17